backtest open positions added. market_mover replayer added

+++ src/strategies/indicators/basic_indicator.py
@@ -1,4 +1,3 @@
-# indicators/obv.py
 import polars as pl
 import talib
 
@@ -7,7 +6,7 @@ from .registry import register
 
 @register("obv", grouped=True)
 def calculate_obv(df: pl.DataFrame) -> pl.DataFrame:
-    close = df["close"].to_numpy()
-    volume = df["volume"].to_numpy().astype("float64")
+    close = df["close"].cast(pl.Float64).to_numpy()
+    volume = df["volume"].cast(pl.Float64).to_numpy().astype("float64")
     obv = talib.OBV(close, volume)
     return df.with_columns(pl.Series("obv", obv))
